Update files structure of page tests
