+++ saleor/graphql/page/tests/mutations/test_page_bulk_delete.py
@@ -1,7 +1,7 @@
 import graphene
 
-from ....page.models import Page
-from ...tests.utils import get_graphql_content
+from .....page.models import Page
+from ....tests.utils import get_graphql_content
 
 
 def test_delete_pages(staff_api_client, page_list, permission_manage_pages):
