+++ saleor/graphql/page/tests/mutations/test_page_types_bulk_delete.py
@@ -1,7 +1,7 @@
 import graphene
 
-from ....page.models import Page
-from ...tests.utils import assert_no_permission, get_graphql_content
+from .....page.models import Page
+from ....tests.utils import assert_no_permission, get_graphql_content
 
 PAGE_TYPE_BULK_DELETE_MUTATION = """
     mutation PageTypeBulkDelete($ids: [ID!]!) {
